feat: qr button
