+++ apps/extension/src/ui/domains/CopyAddress/useCopyAddressModal.ts
@@ -6,9 +6,9 @@ import { atom, useRecoilState } from "recoil"
 
 import { CopyAddressWizardInputs } from "./types"
 
-const copyAddressInputsState = atom<CopyAddressWizardInputs | null>({
+const copyAddressInputsState = atom<CopyAddressWizardInputs>({
   key: "copyAddressInputsState",
-  default: null,
+  default: {},
 })
 
 export const useCopyAddressModal = () => {
@@ -17,12 +17,14 @@ export const useCopyAddressModal = () => {
 
   const open = useCallback(
     (opts: CopyAddressWizardInputs | null) => {
+      // skip wizard for evm addresses
       if (opts && isEthereumAddress(opts.address) && !opts.qr) {
-        // skip wizard for evm addresses
-        copyAddress(opts.address)
+        // if qr isn't explicitely set to false, assume we need to display the copy qr button
+        const onQrClick = opts.qr !== false ? () => open({ ...opts, qr: true }) : undefined
+        copyAddress(opts.address, onQrClick)
       } else {
         // display the wizard
-        setInputs(opts ?? null)
+        setInputs(opts || {})
         innerOpen()
       }
     },
